# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * LinkScope URL shortener and analytics API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type TrendGranularity = typeof TrendGranularity[keyof typeof TrendGranularity];


export const TrendGranularity = {
  hourly: 'hourly',
  daily: 'daily',
  weekly: 'weekly',
  monthly: 'monthly',
} as const;

export interface Link {
  id: number;
  shortCode: string;
  /** Full public short URL (clean /s/{code} path) */
  shortUrl: string;
  /** Legacy /r/{code} redirect path, kept for backward compatibility */
  legacyUrl?: string;
  originalUrl: string;
  /** @nullable */
  title: string | null;
  isEnabled: boolean;
  /** @nullable */
  expiresAt: string | null;
  /** @nullable */
  previewTitle: string | null;
  /** @nullable */
  previewDescription: string | null;
  /** @nullable */
  previewImageUrl: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface LinkStats {
  humanClicks: number;
  uniqueVisitors: number;
  botRequests: number;
  totalRequests: number;
}

export type LinkWithStats = Link & {
  stats: LinkStats;
};

export interface LinkInput {
  /** @minLength 1 */
  originalUrl: string;
  customAlias?: string;
  title?: string;
  /** @nullable */
  expiresAt?: string | null;
  previewTitle?: string;
  previewDescription?: string;
  previewImageUrl?: string;
}

export interface LinkUpdate {
  /** @minLength 1 */
  originalUrl?: string;
  /** @nullable */
  title?: string | null;
  isEnabled?: boolean;
  /** @nullable */
  expiresAt?: string | null;
  /** @nullable */
  previewTitle?: string | null;
  /** @nullable */
  previewDescription?: string | null;
  /** @nullable */
  previewImageUrl?: string | null;
}

export interface QrCode {
  /** base64 PNG data URL */
  dataUrl: string;
  shortUrl: string;
}

export interface AnalyticsOverview {
  totalLinks: number;
  humanClicks: number;
  uniqueVisitors: number;
  botRequests: number;
  facebookRequests: number;
  totalRequests: number;
}

export interface TrendPoint {
  /** ISO timestamp marking the start of the bucket */
  bucket: string;
  humanClicks: number;
  botRequests: number;
}

export interface BreakdownEntry {
  label: string;
  count: number;
  percentage: number;
}

export interface AnalyticsBreakdowns {
  country: BreakdownEntry[];
  city: BreakdownEntry[];
  device: BreakdownEntry[];
  browser: BreakdownEntry[];
  os: BreakdownEntry[];
  referrer: BreakdownEntry[];
  trafficSource: BreakdownEntry[];
  botClassification: BreakdownEntry[];
}

export interface TopLink {
  id: number;
  shortCode: string;
  shortUrl: string;
  /** @nullable */
  title: string | null;
  originalUrl: string;
  humanClicks: number;
  uniqueVisitors: number;
}

export interface HumanClickActivity {
  id: number;
  linkId: number;
  shortCode: string;
  /** @nullable */
  linkTitle: string | null;
  /** @nullable */
  country: string | null;
  /** @nullable */
  city: string | null;
  deviceType: string;
  /** @nullable */
  browser: string | null;
  /** @nullable */
  os: string | null;
  /** @nullable */
  referrer: string | null;
  trafficSource: string;
  clickedAt: string;
}

export interface BotActivity {
  id: number;
  linkId: number;
  shortCode: string;
  /** @nullable */
  linkTitle: string | null;
  classification: string;
  /** @nullable */
  userAgent: string | null;
  requestedAt: string;
}

export type FromDateParameter = string;

export type ToDateParameter = string;

export type FilterLinkIdParameter = number;

export type FilterCountryParameter = string;

export type GetAnalyticsOverviewParams = {
from?: FromDateParameter;
to?: ToDateParameter;
linkId?: FilterLinkIdParameter;
country?: FilterCountryParameter;
};

export type GetAnalyticsTrendParams = {
from?: FromDateParameter;
to?: ToDateParameter;
linkId?: FilterLinkIdParameter;
country?: FilterCountryParameter;
granularity?: TrendGranularity;
};

export type GetAnalyticsBreakdownsParams = {
from?: FromDateParameter;
to?: ToDateParameter;
linkId?: FilterLinkIdParameter;
country?: FilterCountryParameter;
};

export type GetTopLinksParams = {
from?: FromDateParameter;
to?: ToDateParameter;
};

export type GetRecentHumanActivityParams = {
linkId?: FilterLinkIdParameter;
};

export type GetRecentBotActivityParams = {
linkId?: FilterLinkIdParameter;
};