# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: LinkScope URL shortener and analytics API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /analytics/overview:
    get:
      operationId: getAnalyticsOverview
      tags: [analytics]
      summary: Dashboard overview totals (human + bot) for the current user
      parameters:
        - $ref: "#/components/parameters/FromDate"
        - $ref: "#/components/parameters/ToDate"
        - $ref: "#/components/parameters/FilterLinkId"
        - $ref: "#/components/parameters/FilterCountry"
      responses:
        "200":
          description: Overview totals
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalyticsOverview"
  /analytics/trend:
    get:
      operationId: getAnalyticsTrend
      tags: [analytics]
      summary: Human click / bot request time series at a given granularity
      parameters:
        - $ref: "#/components/parameters/FromDate"
        - $ref: "#/components/parameters/ToDate"
        - $ref: "#/components/parameters/FilterLinkId"
        - $ref: "#/components/parameters/FilterCountry"
        - name: granularity
          in: query
          schema:
            $ref: "#/components/schemas/TrendGranularity"
      responses:
        "200":
          description: Trend buckets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TrendPoint"
  /analytics/breakdowns:
    get:
      operationId: getAnalyticsBreakdowns
      tags: [analytics]
      summary: Country, city, device, browser, OS, referrer, traffic source and bot classification breakdowns
      parameters:
        - $ref: "#/components/parameters/FromDate"
        - $ref: "#/components/parameters/ToDate"
        - $ref: "#/components/parameters/FilterLinkId"
        - $ref: "#/components/parameters/FilterCountry"
      responses:
        "200":
          description: Breakdowns
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalyticsBreakdowns"
  /analytics/top-links:
    get:
      operationId: getTopLinks
      tags: [analytics]
      summary: Top performing links by human clicks
      parameters:
        - $ref: "#/components/parameters/FromDate"
        - $ref: "#/components/parameters/ToDate"
      responses:
        "200":
          description: Top links
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TopLink"
  /analytics/recent-activity:
    get:
      operationId: getRecentHumanActivity
      tags: [analytics]
      summary: Recent human click activity
      parameters:
        - $ref: "#/components/parameters/FilterLinkId"
      responses:
        "200":
          description: Recent human clicks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HumanClickActivity"
  /analytics/recent-bot-activity:
    get:
      operationId: getRecentBotActivity
      tags: [analytics]
      summary: Recent bot request activity
      parameters:
        - $ref: "#/components/parameters/FilterLinkId"
      responses:
        "200":
          description: Recent bot requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BotActivity"
components:
  parameters:
...
    FromDate:
      name: from
      in: query
      schema:
        type: string
        format: date-time
    ToDate:
      name: to
      in: query
      schema:
        type: string
        format: date-time
    FilterLinkId:
      name: linkId
      in: query
      schema:
        type: integer
    FilterCountry:
      name: country
      in: query
      schema:
        type: string
  schemas:
...
    TrendGranularity:
      type: string
      enum: [hourly, daily, weekly, monthly]
...
    AnalyticsOverview:
      type: object
      required:
        - totalLinks
        - humanClicks
        - uniqueVisitors
        - botRequests
        - facebookRequests
        - totalRequests
      properties:
        totalLinks:
          type: integer
        humanClicks:
          type: integer
        uniqueVisitors:
          type: integer
        botRequests:
          type: integer
        facebookRequests:
          type: integer
        totalRequests:
          type: integer
    TrendPoint:
      type: object
      required: [bucket, humanClicks, botRequests]
      properties:
        bucket:
          type: string
          description: ISO timestamp marking the start of the bucket
        humanClicks:
          type: integer
        botRequests:
          type: integer
    BreakdownEntry:
      type: object
      required: [label, count, percentage]
      properties:
        label:
          type: string
        count:
          type: integer
        percentage:
          type: number
    AnalyticsBreakdowns:
      type: object
      required:
        - country
        - city
        - device
        - browser
        - os
        - referrer
        - trafficSource
        - botClassification
      properties:
        country:
          type: array
          items:
            $ref: "#/components/schemas/BreakdownEntry"
        city:
          type: array
          items:
            $ref: "#/components/schemas/BreakdownEntry"
        device:
          type: array
          items:
            $ref: "#/components/schemas/BreakdownEntry"
        browser:
          type: array
          items:
            $ref: "#/components/schemas/BreakdownEntry"
        os:
          type: array
          items:
            $ref: "#/components/schemas/BreakdownEntry"
        referrer:
          type: array
          items:
            $ref: "#/components/schemas/BreakdownEntry"
        trafficSource:
          type: array
          items:
            $ref: "#/components/schemas/BreakdownEntry"
        botClassification:
          type: array
          items:
            $ref: "#/components/schemas/BreakdownEntry"
    TopLink:
      type: object
      required: [id, shortCode, shortUrl, title, originalUrl, humanClicks, uniqueVisitors]
      properties:
        id:
          type: integer
        shortCode:
          type: string
        shortUrl:
          type: string
        title:
          type: ["string", "null"]
        originalUrl:
          type: string
        humanClicks:
          type: integer
        uniqueVisitors:
          type: integer
    HumanClickActivity:
      type: object
      required:
        - id
        - linkId
        - shortCode
        - linkTitle
        - country
        - city
        - deviceType
        - browser
        - os
        - referrer
        - trafficSource
        - clickedAt
      properties:
        id:
          type: integer
        linkId:
          type: integer
        shortCode:
          type: string
        linkTitle:
          type: ["string", "null"]
        country:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        deviceType:
          type: string
        browser:
          type: ["string", "null"]
        os:
          type: ["string", "null"]
        referrer:
          type: ["string", "null"]
        trafficSource:
          type: string
        clickedAt:
          type: string
          format: date-time
    BotActivity:
      type: object
      required:
        - id
        - linkId
        - shortCode
        - linkTitle
        - classification
        - userAgent
        - requestedAt
      properties:
        id:
          type: integer
        linkId:
          type: integer
        shortCode:
          type: string
        linkTitle:
          type: ["string", "null"]
        classification:
          type: string
        userAgent:
          type: ["string", "null"]
        requestedAt:
          type: string
          format: date-time