# lib/api-zod/src/generated/types/analyticsOverview.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * LinkScope URL shortener and analytics API
 * OpenAPI spec version: 0.1.0
 */

export interface AnalyticsOverview {
  totalLinks: number;
  humanClicks: number;
  uniqueVisitors: number;
  botRequests: number;
  facebookRequests: number;
  totalRequests: number;
}