# lib/api-zod/src/generated/types/trendGranularity.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * LinkScope URL shortener and analytics API
 * OpenAPI spec version: 0.1.0
 */

export type TrendGranularity = typeof TrendGranularity[keyof typeof TrendGranularity];


export const TrendGranularity = {
  hourly: 'hourly',
  daily: 'daily',
  weekly: 'weekly',
  monthly: 'monthly',
} as const;