# lib/api-zod/src/generated/types/trendPoint.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * LinkScope URL shortener and analytics API
 * OpenAPI spec version: 0.1.0
 */

export interface TrendPoint {
  /** ISO timestamp marking the start of the bucket */
  bucket: string;
  humanClicks: number;
  botRequests: number;
}